+++ src/o/empty_list.js
@@ -1,0 +1,77 @@
+const AbstractList = require("./abstract_list");
+
+// `List` extends `AbstractList` and is
+// the primary object in Bubblescript and
+// is the programatic representation of a
+// list. e.g. `(1 2 3)`
+class List extends AbstractList {
+
+  // `List.emptyList` provides an instance
+  // of `EmptyList`, which terminates all
+  // lists.
+  static get emptyList() { return emptyList; }
+
+  // `List.make` makes/creates a new list.
+  // `List.make(1, 2, 3)`
+  static make(...elements) {
+    return List._make(elements);
+  }
+
+  static _make(elementsArray, currentLinkedList=emptyList) {
+    if (elementsArray.length < 1)
+      return currentLinkedList;
+    return List._make(elementsArray,
+      new List(elementsArray.pop(),
+        currentLinkedList));
+  }
+
+  // Create a list.
+  constructor(o, oo=emptyList) {
+    super(o, oo);
+  }
+
+  push(element) {
+    return new List(element, this);
+  }
+
+  toString() {
+    return "(" + this._toString() + ")";
+  }
+
+  toStringJoin(accumulatedString, formattedElement) {
+    return accumulatedString + " " + formattedElement;
+  };
+
+  toVector() {
+    return this.reduce((vector, o) => {
+      return vector.push(o); },
+      Vector.emptyVector);
+  }
+
+  each(fn) {
+    let result;
+    try {
+      result = fn(this.peek());
+    } catch (o) {
+      if (o instanceof MacroExpanded) {
+        let expanded = o.expanded;
+        this.o  = expanded.first;
+        this.oo = this.rest.conj(expanded.rest.invert());
+        return this.each(fn);
+      } else {
+        throw o;
+      }
+    }
+    if (this.pop().isEmpty) return result;
+    return this.pop().each(fn);
+  }
+
+}
+
+class EmptyList extends List {
+  get isEmpty() { return true; }
+}
+
+emptyList = new EmptyList()
+
+module.exports = List;
